+++ src/main/java/net/finmath/montecarlo/assetderivativevaluation/models/VarianceGammaModel.java
@@ -4,6 +4,7 @@ import java.util.Map;
 
 import net.finmath.exception.CalculationException;
 import net.finmath.marketdata.model.curves.DiscountCurve;
+import net.finmath.modelling.descriptor.VarianceGammaModelDescriptor;
 import net.finmath.montecarlo.model.AbstractProcessModel;
 import net.finmath.montecarlo.model.ProcessModel;
 import net.finmath.stochastic.RandomVariable;
@@ -42,14 +43,28 @@ public class VarianceGammaModel extends AbstractProcessModel {
 	private final double nu;
 
 	/**
+	 * Create the model from a descriptor.
+	 *
+	 * @param descriptor A descriptor of the model.
+	 */
+	public VarianceGammaModel(VarianceGammaModelDescriptor descriptor) {
+		this(descriptor.getInitialValue(),
+				descriptor.getDiscountCurveForForwardRate(),
+				descriptor.getDiscountCurveForDiscountRate(),
+				descriptor.getSigma(),
+				descriptor.getTheta(),
+				descriptor.getNu());
+	}
+
+	/**
 	 * Construct a Variance Gamma model with discount curves for the forward price (i.e. repo rate minus dividend yield) and for discounting.
 	 *
 	 * @param initialValue \( S_{0} \) - spot - initial value of S
 	 * @param discountCurveForForwardRate The curve specifying \( t \mapsto exp(- r^{\text{c}}(t) \cdot t) \) - with \( r^{\text{c}}(t) \) the risk free rate
 	 * @param discountCurveForDiscountRate The curve specifying \( t \mapsto exp(- r^{\text{d}}(t) \cdot t) \) - with \( r^{\text{d}}(t) \) the discount rate
-	 * @param sigma The parameter \( \sigma \)
-	 * @param theta The parameter \( \theta \)
-	 * @param nu The parameter \( \nu \)
+	 * @param sigma The parameter \( \sigma \).
+	 * @param theta The parameter \( \theta \).
+	 * @param nu The parameter \( \nu \).
 	 */
 	public VarianceGammaModel(final double initialValue, final DiscountCurve discountCurveForForwardRate,
 			final DiscountCurve discountCurveForDiscountRate, final double sigma, final double theta, final double nu) {
@@ -70,9 +85,9 @@ public class VarianceGammaModel extends AbstractProcessModel {
 	 * @param initialValue \( S_{0} \) - spot - initial value of S
 	 * @param riskFreeRate The constant risk free rate for the drift (repo rate of the underlying).
 	 * @param discountRate The constant rate used for discounting.
-	 * @param sigma The parameter \( \sigma \)
-	 * @param theta The parameter \( \theta \)
-	 * @param nu The parameter \( \nu \)
+	 * @param sigma The parameter \( \sigma \).
+	 * @param theta The parameter \( \theta \).
+	 * @param nu The parameter \( \nu \).
 	 */
 	public VarianceGammaModel(final double initialValue, final double riskFreeRate, final double discountRate, final double sigma, final double theta,
 			final double nu) {
@@ -93,9 +108,9 @@ public class VarianceGammaModel extends AbstractProcessModel {
 	 *
 	 * @param initialValue \( S_{0} \) - spot - initial value of S
 	 * @param riskFreeRate The constant risk free rate for the drift (repo rate of the underlying).
-	 * @param sigma The parameter \( \sigma \)
-	 * @param theta The parameter \( \theta \)
-	 * @param nu The parameter \( \nu \)
+	 * @param sigma The parameter \( \sigma \).
+	 * @param theta The parameter \( \theta \).
+	 * @param nu The parameter \( \nu \).
 	 */
 	public VarianceGammaModel(final double initialValue, final double riskFreeRate, final double sigma, final double theta, final double nu) {
 		this(initialValue,riskFreeRate,riskFreeRate,sigma,theta,nu);